### 2025-04-17 #8068 @editorialbot
[PRE REVIEW]: PowerFactory-Tools: A toolbox to facilitate the control of DIgSILENT PowerFactory, written in Python
**Submitting author:** @SebastianDD (<a href="http://orcid.org/0000-0002-3940-3398">Sebastian Krahmer</a>)
**Repository:** https://github.com/ieeh-tu-dresden/powerfactory-tools
**Branch with paper.md** (empty if default branch): SebastianDD/166-docs-add-JOSS-paper
**Version:** v3.1.0
