### 2025-04-22 74c0f3a
docs: Improve paper title
