### 2025-05-02 105c942
docs: add Social Media Preview Card Fixes #187
