### 2025-05-05 00b6701
docs: add Social Media Preview Card (#306)
Fixes #187
