### 2025-06-25 #8068 @robinroche
@kyleniemeyer: before answering, I need to make sure I have the required software.
@SebastianDD: what are the requirements on the PowerFactory side? I only have an educational version, which does not include scripting features, as far as I know.
