### 2025-07-16 c4a34cf
feat: add run_short_circuit_calculation() Fixes #287
