### 2025-07-16 6cc2c0f
feat: exporter - make plausibility check of topology case optional. fixes #309
