+++ src/powerfactory_tools/versions/pf2024/exporter/exporter.py
@@ -159,6 +159,7 @@ class PowerFactoryExporterProcess(multiprocessing.Process):
         steadystate_case_name: str | None = None,
         study_case_names: list[str] | None = None,
         element_specific_attrs: dict[PFClassId, Sequence[str | dict]] | None = None,
+        plausibility_check: bool = True,
     ) -> None:
         super().__init__()
         self.export_path = export_path
@@ -176,6 +177,7 @@ class PowerFactoryExporterProcess(multiprocessing.Process):
         self.steadystate_case_name = steadystate_case_name
         self.study_case_names = study_case_names
         self.element_specific_attrs = element_specific_attrs
+        self.plausibility_check = plausibility_check
 
     def run(self) -> None:
         pfe = PowerFactoryExporter(
@@ -196,6 +198,7 @@ class PowerFactoryExporterProcess(multiprocessing.Process):
             topology_case_name=self.topology_case_name,
             steadystate_case_name=self.steadystate_case_name,
             study_case_names=self.study_case_names,
+            plausibility_check=self.plausibility_check,
         )
 
 
@@ -245,6 +248,7 @@ class PowerFactoryExporter:
         topology_case_name: str | None = None,
         steadystate_case_name: str | None = None,
         study_case_names: list[str] | None = None,
+        plausibility_check: bool = True,
     ) -> None:
         """Export grid topology, topology_case and steadystate_case to json files.
 
@@ -258,6 +262,7 @@ class PowerFactoryExporter:
             topology_case_name {str} -- the chosen file name for related 'topology_case' data (default: {None})
             steadystate_case_name {str} -- the chosen file name for related 'steadystate_case' data (default: {None})
             study_case_names {list[str]} -- a list of study cases to export (default: {None})
+            plausibility_check {bool} -- flag to turn on/off the plausibility check of the exported topology_case (default: {True})
         """
         if study_case_names is not None:
             self.export_study_cases(
@@ -266,6 +271,7 @@ class PowerFactoryExporter:
                 topology_name=topology_name,
                 topology_case_name=topology_case_name,
                 steadystate_case_name=steadystate_case_name,
+                plausibility_check=plausibility_check,
             )
         else:
             act_sc = self.pfi.app.GetActiveStudyCase()
@@ -276,6 +282,7 @@ class PowerFactoryExporter:
                     topology_name=topology_name,
                     topology_case_name=topology_case_name,
                     steadystate_case_name=steadystate_case_name,
+                    plausibility_check=plausibility_check,
                 )
             else:
                 msg = "Could not export. There is neither a study case defined nor is one activated"
@@ -289,6 +296,7 @@ class PowerFactoryExporter:
         topology_name: str | None,
         topology_case_name: str | None,
         steadystate_case_name: str | None,
+        plausibility_check: bool = True,
     ) -> None:
         for study_case_name in study_case_names:
             study_case = self.pfi.study_case(study_case_name)
@@ -300,6 +308,7 @@ class PowerFactoryExporter:
                     topology_name=topology_name,
                     topology_case_name=topology_case_name,
                     steadystate_case_name=steadystate_case_name,
+                    plausibility_check=plausibility_check,
                 )
             else:
                 loguru.logger.warning(
@@ -315,6 +324,7 @@ class PowerFactoryExporter:
         topology_name: str | None,
         topology_case_name: str | None,
         steadystate_case_name: str | None,
+        plausibility_check: bool = True,
     ) -> None:
         grids = self.pfi.independent_grids(calc_relevant=True)
 
@@ -332,7 +342,12 @@ class PowerFactoryExporter:
 
[…]
