### 2025-07-16 ab7163b
fix: pf class id of external grid. fixes #311
