### 2025-07-16 4a91c92
fix: mypy issues and code style
