+++ pyproject.toml
@@ -73,6 +73,7 @@
       "D401",
       "E402",
       "E501",
+      "G004",
     ]
     select = ["ALL"]
 
