### 2025-07-16 c852246
fix: ruff
