### 2025-07-17 a8320f0
docs: be more precise in readme
