### 2025-07-17 6e9a89f
feat: added powerfactory_ini_name, powerfactory_user_password, and plausibility_check parameters to exporter interfaces and methods (#310)
Fixes #308
Fixes #309
Fixes #311
