+++ src/powerfactory_tools/versions/pf2024/exporter/exporter.py
@@ -146,9 +146,11 @@ class PowerFactoryExporterProcess(multiprocessing.Process):
         *,
         project_name: str,
         export_path: pathlib.Path,
-        powerfactory_user_profile: str = "",
+        powerfactory_ini_name: str | None = None,
         powerfactory_path: pathlib.Path = DEFAULT_POWERFACTORY_PATH,
         powerfactory_service_pack: int | None = None,
+        powerfactory_user_profile: str = "",
+        powerfactory_user_password: str | None = None,
         python_version: ValidPythonVersion = DEFAULT_PYTHON_VERSION,
         logging_level: int = logging.DEBUG,
         log_file_path: pathlib.Path | None = None,
@@ -157,13 +159,16 @@ class PowerFactoryExporterProcess(multiprocessing.Process):
         steadystate_case_name: str | None = None,
         study_case_names: list[str] | None = None,
         element_specific_attrs: dict[PFClassId, Sequence[str | dict]] | None = None,
+        plausibility_check: bool = True,
     ) -> None:
         super().__init__()
         self.export_path = export_path
         self.project_name = project_name
-        self.powerfactory_user_profile = powerfactory_user_profile
+        self.powerfactory_ini_name = powerfactory_ini_name
         self.powerfactory_path = powerfactory_path
         self.powerfactory_service_pack = powerfactory_service_pack
+        self.powerfactory_user_profile = powerfactory_user_profile
+        self.powerfactory_user_password = powerfactory_user_password
         self.python_version = python_version
         self.logging_level = logging_level
         self.log_file_path = log_file_path
@@ -172,13 +177,16 @@ class PowerFactoryExporterProcess(multiprocessing.Process):
         self.steadystate_case_name = steadystate_case_name
         self.study_case_names = study_case_names
         self.element_specific_attrs = element_specific_attrs
+        self.plausibility_check = plausibility_check
 
     def run(self) -> None:
         pfe = PowerFactoryExporter(
             project_name=self.project_name,
-            powerfactory_user_profile=self.powerfactory_user_profile,
+            powerfactory_ini_name=self.powerfactory_ini_name,
             powerfactory_path=self.powerfactory_path,
             powerfactory_service_pack=self.powerfactory_service_pack,
+            powerfactory_user_profile=self.powerfactory_user_profile,
+            powerfactory_user_password=self.powerfactory_user_password,
             python_version=self.python_version,
             logging_level=self.logging_level,
             log_file_path=self.log_file_path,
@@ -190,15 +198,18 @@ class PowerFactoryExporterProcess(multiprocessing.Process):
             topology_case_name=self.topology_case_name,
             steadystate_case_name=self.steadystate_case_name,
             study_case_names=self.study_case_names,
+            plausibility_check=self.plausibility_check,
         )
 
 
 @pydantic.dataclasses.dataclass
 class PowerFactoryExporter:
     project_name: str
-    powerfactory_user_profile: str = ""
+    powerfactory_ini_name: str | None = None
     powerfactory_path: pathlib.Path = DEFAULT_POWERFACTORY_PATH
     powerfactory_service_pack: int | None = None
+    powerfactory_user_profile: str = ""
+    powerfactory_user_password: str | None = None
     python_version: ValidPythonVersion = DEFAULT_PYTHON_VERSION
     logging_level: int = logging.DEBUG
     log_file_path: pathlib.Path | None = None
@@ -207,9 +218,11 @@ class PowerFactoryExporter:
     def __post_init__(self) -> None:
         self.pfi = PowerFactoryInterface(
             project_name=self.project_name,
-            powerfactory_user_profile=self.powerfactory_user_profile,
[…]
