### 2025-07-17 cb752a0
feat: handle dataframe for  data export/import functions Fixes #283
