+++ src/powerfactory_tools/utils/io.py
@@ -3,14 +3,14 @@
 # :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2025.
 # :license: BSD 3-Clause
 
+# pyright: reportPossiblyUnboundVariable=false
+
 from __future__ import annotations
 
 import abc
-import csv
+import ast
 import enum
-import json
 import pathlib
-import pickle
 import typing as t
 
 import loguru
@@ -47,13 +47,53 @@ class FileType(enum.Enum):
         return value in cls.values()
 
 
+def _format_dict(data: dict[str, PrimitiveType]) -> dict[str, PrimitiveType]:
+    # Convert dictionary to list of dictionaries
+    max_length = max(len(v) if isinstance(v, list) else 1 for v in data.values())
+    # Convert non-list values to lists with repeated values and pad shorter lists with None
+    return {
+        key: (v + [None] * (max_length - len(v))) if isinstance(v, list) else ([v] + [None] * (max_length - 1))
+        for key, v in data.items()
+    }
+
+
+def convert_dataframe_to_dict(dataframe: pd.DataFrame) -> dict[str, PrimitiveType]:
+    """Convert a pandas DataFrame to a dict."""
+    # Drop NaN values from each column
+    data = {col: dataframe[col].dropna().tolist() for col in dataframe.columns}
+    # Unmap lists with a single value back to a single value
+    for key, value in data.items():
+        if all(v == value[0] for v in value):
+            data[key] = value[0]
+
+    return data
+
+
+def convert_dict_to_dataframe(data: dict[str, PrimitiveType] | pd.DataFrame) -> pd.DataFrame:
+    """Convert dict to a pandas DataFrame."""
+    if isinstance(data, dict):
+        padded_data = _format_dict(data)
+        return pd.DataFrame.from_dict(padded_data)
+
+    return data
+
+
+def convert_str_lists_to_lists(df: pd.DataFrame, columns: cabc.Sequence[str]) -> pd.DataFrame:
+    """Convert string representations of lists in specified columns to actual lists."""
+    for col in columns:
+        df[col] = df[col].apply(
+            lambda x: ast.literal_eval(x) if isinstance(x, str) and x.startswith("[") and x.endswith("]") else x,
+        )
+    return df
+
+
 @pydantic.dataclasses.dataclass
 class ExportHandler(abc.ABC):
     directory_path: pathlib.Path
 
     def export_user_data(
         self,
-        data: dict[str, PrimitiveType],
+        data: dict[str, PrimitiveType] | pd.DataFrame,
         /,
         *,
         file_type: FileType,
@@ -102,20 +142,11 @@ class ExportHandler(abc.ABC):
         raise NotImplementedError(msg)
 
     @staticmethod
-    def _format_dict(data: dict[str, PrimitiveType]) -> dict[str, PrimitiveType]:
-        # Convert dictionary to list of dictionaries
-        max_length = max(len(v) if isinstance(v, list) else 1 for v in data.values())
-        # Convert non-list values to lists with repeated values and pad shorter lists with None
-        return {
-            key: (v + [None] * (max_length - len(v))) if isinstance(v, list) else ([v] + [None] * (max_length - 1))
-            for key, v in data.items()
-        }
-
-    def to_json(self, file_path: pathlib.Path, /, *, data: dict[str, PrimitiveType], indent: int = 2) -> bool:
-        padded_data = self._format_dict(data)
+    def to_json(file_path: pathlib.Path, /, *, data: dict[str, PrimitiveType] | pd.DataFrame, indent: int = 2) -> bool:
+        dataframe = convert_dict_to_dataframe(data)
         try:
-            with pathlib.Path(file_path).open("w+", encoding="utf-8") as file_handle:
[…]
