+++ tests/test_io.py
@@ -1,24 +1,42 @@
 from contextlib import nullcontext as does_not_raise
 
+import pandas as pd
 import pytest
 
 from powerfactory_tools.utils.io import FileType
 from powerfactory_tools.utils.io import ImportHandler
+from powerfactory_tools.utils.io import convert_dataframe_to_dict
+from powerfactory_tools.utils.io import convert_dict_to_dataframe
+from powerfactory_tools.utils.io import convert_str_lists_to_lists
 from powerfactory_tools.versions.pf2024.utils.io import ExportHandler
 
-test_data = {
+test_dict1 = {
     "name": ["Alice", "Bob", "Charlie"],
     "age": [25, 30, 35],
     "city": ["New York", "Los Angeles"],
 }
+test_df1 = pd.DataFrame(
+    {
+        "name": ["Alice", "Bob", "Charlie"],
+        "age": [25, 30, 35],
+        "city": ["New York", "Los Angeles", None],  # Fill missing value with None/NaN
+    },
+)
 
-test_data2 = {
+test_dict2 = {
     "name": "Alice",
     "age": 25,
     "city": ["New York", "Los Angeles"],
 }
+test_df2 = pd.DataFrame(
+    {
+        "name": ["Alice"],
+        "age": [25],
+        "city": [["New York", "Los Angeles"]],  # List as a single cell
+    },
+)
 
-test_data3 = {
+test_dict3 = {
     "name": "Alice",
     "age": 25,
     "city": "New York",
@@ -34,15 +52,21 @@ class TestExportImportData:
     @pytest.mark.parametrize(
         ("data", "file_type", "expectation"),
         [
-            (test_data, FileType.CSV, does_not_raise()),
-            (test_data, FileType.JSON, does_not_raise()),
-            (test_data, FileType.FEATHER, does_not_raise()),
-            (test_data2, FileType.CSV, does_not_raise()),
-            (test_data2, FileType.JSON, does_not_raise()),
-            (test_data2, FileType.FEATHER, does_not_raise()),
-            (test_data3, FileType.CSV, does_not_raise()),
-            (test_data3, FileType.JSON, does_not_raise()),
-            (test_data3, FileType.FEATHER, does_not_raise()),
+            (test_dict1, FileType.CSV, does_not_raise()),
+            (test_dict1, FileType.JSON, does_not_raise()),
+            (test_dict1, FileType.FEATHER, does_not_raise()),
+            (test_dict2, FileType.CSV, does_not_raise()),
+            (test_dict2, FileType.JSON, does_not_raise()),
+            (test_dict2, FileType.FEATHER, does_not_raise()),
+            (test_dict3, FileType.CSV, does_not_raise()),
+            (test_dict3, FileType.JSON, does_not_raise()),
+            (test_dict3, FileType.FEATHER, does_not_raise()),
+            (test_df1, FileType.CSV, does_not_raise()),
+            (test_df1, FileType.JSON, does_not_raise()),
+            (test_df1, FileType.FEATHER, does_not_raise()),
+            (test_df2, FileType.CSV, does_not_raise()),
+            (test_df2, FileType.JSON, does_not_raise()),
+            (test_df2, FileType.FEATHER, does_not_raise()),
         ],
     )
     def test_export_import_user_data(
@@ -62,5 +86,20 @@ class TestExportImportData:
             file_name = f"{file_name}{file_type.value}"
             full_file_path = fake_path / file_name
             ih = ImportHandler(file_path=full_file_path)
-            imported_data = ih.import_user_data()
-            assert data == imported_data
+            imported_dataframe = ih.import_user_data()
+            if imported_dataframe is None:
+                pytest.fail("Import returned None, expected a DataFrame.")
[…]
