### 2025-07-17 52fdddb
fix: remove commented-out code
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
