### 2025-07-17 2195704
fix: no binary mode wb for json export
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
