### 2025-07-17 50783f1
fix: remove doubled code in test_io
