### 2025-07-18 1a85025
feat: add run_short_circuit_calculation() for pf version 2024 (#307)
Fixes #287
