+++ src/powerfactory_tools/versions/pf2024/interface.py
@@ -36,6 +36,10 @@ from powerfactory_tools.versions.pf2024.types import NetworkExtendedCalcType
 from powerfactory_tools.versions.pf2024.types import PFClassId
 from powerfactory_tools.versions.pf2024.types import PowerFactoryTypes as PFTypes
 from powerfactory_tools.versions.pf2024.types import ResultExportMode
+from powerfactory_tools.versions.pf2024.types import ShortCircuitCalcType
+from powerfactory_tools.versions.pf2024.types import ShortCircuitCurrentType
+from powerfactory_tools.versions.pf2024.types import ShortCircuitFaultType
+from powerfactory_tools.versions.pf2024.types import ShortCircuitPosition
 from powerfactory_tools.versions.pf2024.types import TimeSimulationNetworkCalcType
 from powerfactory_tools.versions.pf2024.types import TimeSimulationType
 from powerfactory_tools.versions.pf2024.types import UnitSystem
@@ -365,11 +369,11 @@ class PowerFactoryInterface:
     def compile_powerfactory_data(self, grid: PFTypes.Grid) -> PowerFactoryData:
         """Read out all relevant data from PowerFactory for a given grid and store as typed dataclass PowerFactroyData.
 
-        Args:
-            grid (PFTypes.Grid): the grid object to be read out
+        Arguments:
+            grid {PFTypes.Grid} -- the grid object to be read out
 
         Returns:
-            PowerFactoryData: a dataclass containing typed lists with all relevant data from PowerFactory
+            {PowerFactoryData} -- a dataclass containing typed lists with all relevant data from PowerFactory
         """
         grid_name = grid.loc_name
         loguru.logger.debug("Compiling data from PowerFactory for grid {grid_name}...", grid_name=grid_name)
@@ -409,10 +413,10 @@ class PowerFactoryInterface:
     ) -> None:
         """Adds variables to a result object for a given list of elements.
 
-        Args:
-            result (PFTypes.Result): the result object to be written to
-            elements (Sequence[PFTypes.DataObject]): list of elements for which variables has to be added to the result object
-            variables (Sequence[str]): list of variables (string identifiers) to be added for each element
+        Arguments:
+            result {PFTypes.Result} -- the result object to be written to
+            elements {Sequence[PFTypes.DataObject]} -- list of elements for which variables has to be added to the result object
+            variables {Sequence[str]} -- list of variables (string identifiers) to be added for each element
         """
         loguru.logger.debug("Set Variables for result object {result_name} ...", result_name=result.loc_name)
         for elm in elements:
@@ -426,11 +430,11 @@ class PowerFactoryInterface:
         In case of load flow calculation, these variable monitors can be written (kind of making visible)
         as result variables to the result itself.
 
-        Args:
-            result (PFTypes.Result): the result object to be written
+        Arguments:
+            result {PFTypes.Result} -- the result object to be written
 
         Raises:
-            RuntimeError: if the writing of the variable monitors fails
+            {RuntimeError} -- if the writing of the variable monitors fails
         """
         loguru.logger.debug(
             f"Write all in the result object {result.loc_name} containing variable monitors (IntMon) as result variables ...",
@@ -1499,13 +1503,14 @@ class PowerFactoryInterface:
     ) -> Sequence[PFTypes.Terminal]:
         """Get all slack terminals of the specified grid based on the ExternalGrid objects.
 
-        Args:
-            grid_name (str):  Name of grid to be accessed.
-            calc_relevant (bool, optional): Flag, if only calc relevant (active) external_grids should be accessed. Defaults to False.
[…]
