### 2025-07-18 0cb8dff
feat: refactor whole io handling, add polars dataframes
