+++ src/powerfactory_tools/str_constants.py
@@ -1,0 +1,7 @@
+# :author: Sebastian Krahmer <sebastian.krahmer@tu-dresden.de>
+# :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2025.
+# :license: BSD 3-Clause
+
+NAME_SEPARATOR = "__"
+PATH_SEPARATOR = "/"
+STRING_SEPARATOR = " || "
