+++ tests/test_io.py
@@ -1,45 +1,61 @@
 from contextlib import nullcontext as does_not_raise
 
 import pandas as pd
+import polars as pl
+import json
 import pytest
 
 from powerfactory_tools.utils.io import FileType
-from powerfactory_tools.utils.io import ImportHandler
-from powerfactory_tools.utils.io import convert_dataframe_to_dict
-from powerfactory_tools.utils.io import convert_dict_to_dataframe
-from powerfactory_tools.utils.io import convert_str_lists_to_lists
-from powerfactory_tools.versions.pf2024.utils.io import ExportHandler
+from powerfactory_tools.utils.io import PandasIoHandler
+from powerfactory_tools.utils.io import PolarsIoHandler
 
 test_dict1 = {
     "name": ["Alice", "Bob", "Charlie"],
     "age": [25, 30, 35],
-    "city": ["New York", "Los Angeles"],
+    "city": ["New York", "Los Angeles", "nn"],
 }
-test_df1 = pd.DataFrame(
+test_pd_df1 = pd.DataFrame(
     {
         "name": ["Alice", "Bob", "Charlie"],
         "age": [25, 30, 35],
         "city": ["New York", "Los Angeles", None],  # Fill missing value with None/NaN
     },
 )
+test_pl_df1 = pl.from_pandas(test_pd_df1)
 
 test_dict2 = {
     "name": "Alice",
     "age": 25,
     "city": ["New York", "Los Angeles"],
 }
-test_df2 = pd.DataFrame(
+test_pd_df2 = pd.DataFrame(
     {
         "name": ["Alice"],
         "age": [25],
         "city": [["New York", "Los Angeles"]],  # List as a single cell
     },
 )
+test_pl_df2 = pl.from_pandas(test_pd_df2)
 
 test_dict3 = {
     "name": "Alice",
     "age": 25,
     "city": "New York",
+}
+
+test_nested_dict1 = {
+    "key_name": {
+        "Uabs1": {
+            "value": 2.3,
+            "unit": "kV",
+        },
+    },
+    "key_name2": {
+        "Uabs1": {
+            "value": 30.3,
+            "unit": "kV",
+        },
+    },
 }
 
 
@@ -48,7 +64,7 @@ def fake_path(tmp_path):
     return tmp_path
 
 
-class TestExportImportData:
+class TestExportImportDataPandas1:
     @pytest.mark.parametrize(
         ("data", "file_type", "expectation"),
         [
@@ -61,12 +77,15 @@ class TestExportImportData:
             (test_dict3, FileType.CSV, does_not_raise()),
             (test_dict3, FileType.JSON, does_not_raise()),
             (test_dict3, FileType.FEATHER, does_not_raise()),
-            (test_df1, FileType.CSV, does_not_raise()),
-            (test_df1, FileType.JSON, does_not_raise()),
-            (test_df1, FileType.FEATHER, does_not_raise()),
-            (test_df2, FileType.CSV, does_not_raise()),
-            (test_df2, FileType.JSON, does_not_raise()),
-            (test_df2, FileType.FEATHER, does_not_raise()),
+            (test_pd_df1, FileType.CSV, does_not_raise()),
+            (test_pd_df1, FileType.JSON, does_not_raise()),
+            (test_pd_df1, FileType.FEATHER, does_not_raise()),
+            (test_pd_df2, FileType.CSV, does_not_raise()),
+            (test_pd_df2, FileType.JSON, does_not_raise()),
+            (test_pd_df2, FileType.FEATHER, does_not_raise()),
+            (test_nested_dict1, FileType.CSV, does_not_raise()),
+            (test_nested_dict1, FileType.JSON, does_not_raise()),
[…]
