### 2025-07-21 ac5794f
refactor: replace ExportHandler with PandasIoHandler and finsih tests
