### 2025-07-21 e685505
fix: update FEATHER file type extension for clarity
