### 2025-07-28 b333432
fix: review comments, add exclude options for pyright to pyproject.toml
