### 2025-07-28 8e3bf58
feat: handle dataframe for data export/import functions (#312)
Fixes #283
