+++ src/powerfactory_tools/versions/pf2022/exporter/exporter.py
@@ -66,10 +66,10 @@ from psdm.topology_case.case import Case as TopologyCase
 from psdm.topology_case.element_state import ElementState
 
 from powerfactory_tools.__version__ import VERSION
+from powerfactory_tools.str_constants import NAME_SEPARATOR
+from powerfactory_tools.str_constants import STRING_SEPARATOR
 from powerfactory_tools.utils.io import FileType
 from powerfactory_tools.versions.pf2022.constants import DEFAULT_PHASE_QUANTITY
-from powerfactory_tools.versions.pf2022.constants import NAME_SEPARATOR
-from powerfactory_tools.versions.pf2022.constants import STRING_SEPARATOR
 from powerfactory_tools.versions.pf2022.constants import DecimalDigits
 from powerfactory_tools.versions.pf2022.constants import Exponents
 from powerfactory_tools.versions.pf2022.exporter.load_power import ConsolidatedLoadPhaseConnectionType
@@ -448,14 +448,16 @@ class PowerFactoryExporter:
         grid_name = data.grid_name.replace(" ", "-")
         project_name = data.project_name.replace(" ", "-")
         date = data.date
-        pf_version_data = AttributeData(
-            name="PowerFactoryVersion",
-            value=(
-                POWERFACTORY_VERSION
-                if self.pfi.powerfactory_service_pack is None
-                else POWERFACTORY_VERSION + STRING_SEPARATOR + "SP" + str(self.pfi.powerfactory_service_pack)
+        pf_version_data = tuple(
+            AttributeData(
+                name="PowerFactoryVersion",
+                value=(
+                    POWERFACTORY_VERSION
+                    if self.pfi.powerfactory_service_pack is None
+                    else POWERFACTORY_VERSION + STRING_SEPARATOR + "SP" + str(self.pfi.powerfactory_service_pack)
+                ),
+                description="The version of PowerFactory used for export.",
             ),
-            description="The version of PowerFactory used for export.",
         )
 
         return Meta(
@@ -465,7 +467,7 @@ class PowerFactoryExporter:
             project=project_name,
             sign_convention=SignConvention.PASSIVE,
             creator=f"powerfactory-tools @ version {VERSION}",
-            optional_data=[pf_version_data],
+            optional_data=pf_version_data,
         )
 
     def create_topology(
@@ -1151,7 +1153,7 @@ class PowerFactoryExporter:
                 tap_side=tap_side,
                 description=description,
                 phase_technology_type=ph_technology,
-                windings=[wh, wl],
+                windings=(wh, wl),
                 optional_data=extra_meta_data,
             )
 
@@ -4145,13 +4147,14 @@ class PowerFactoryExporter:
 
         In case of the occurence of DataObject as value (return type) of a requested attribute: If the grid_name is given, the DataObject is converted to its unique_name + class_name , otherwise the full name is used.
 
-        Args:
-            element (PFTypes.DataObject): the element of interest
-            element_specific_attrs (dict[PFClassId, set[str]]): a dictionary with PFClassIds as keys and a set of attribute names as value
-            grid_name (str | None, optional): the name of the grid related to the element, relevant if converting a PFTypes.DataObject. Defaults to None.
+        Arguments:
+            element {PFTypes.DataObject} -- the element of interest
+            element_specific_attrs {dict[PFClassId, set[str]]} -- a dictionary with PFClassIds as keys and a set of attribute names as value
+        Keyword Arguments:
+            grid_name {str | None} -- the name of the grid related to the element, relevant if converting a PFTypes.DataObject. (default: {None})
 
         Returns:
-            Sequence[AttributeData] | None: list of AttributeData or None if no attributes have been defined for this element type
[…]
