+++ src/powerfactory_tools/utils/__init__.py
@@ -2,12 +2,15 @@
 # :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2025.
 # :license: BSD 3-Clause
 
-from powerfactory_tools.utils.io import ExportHandler
+
+from powerfactory_tools.utils.io import BaseIoHandler
 from powerfactory_tools.utils.io import FileType
-from powerfactory_tools.utils.io import ImportHandler
+from powerfactory_tools.utils.io import PandasIoHandler
+from powerfactory_tools.utils.io import PolarsIoHandler
 
 __all__ = [
-    "ExportHandler",
+    "BaseIoHandler",
     "FileType",
-    "ImportHandler",
+    "PandasIoHandler",
+    "PolarsIoHandler",
 ]
