+++ src/powerfactory_tools/utils/io.py
@@ -3,23 +3,19 @@
 # :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2025.
 # :license: BSD 3-Clause
 
+
 from __future__ import annotations
 
-import abc
-import csv
+import ast
+import datetime as dt
 import enum
 import json
 import pathlib
-import pickle
 import typing as t
 
 import loguru
-import pydantic
 
-try:
-    import pandas as pd
-except ModuleNotFoundError:
-    loguru.logger.error("Missing optional dependency 'pandas'. Use pip or conda to install pandas.")
+from powerfactory_tools.str_constants import NAME_SEPARATOR
 
 if t.TYPE_CHECKING:
     import collections.abc as cabc
@@ -30,9 +26,9 @@ if t.TYPE_CHECKING:
 
 
 class FileType(enum.Enum):
-    FEATHER = ".feather"  # exchange format for dataframes using pyarrow internally
-    DAT = ".dat"  # e.g. for COMTRADE
     CSV = ".csv"
+    DAT = ".dat"  # e.g. for COMTRADE
+    FEATHER = ".arrow"  # exchange format for dataframes using pyarrow (Feather (V2) is a full set of the Arrow IPC file format)
     JSON = ".json"
     PICKLE = ".pkl"
     RAW = ".raw"  # e.g. for PSSPLT_VERSION_2
@@ -47,36 +43,129 @@ class FileType(enum.Enum):
         return value in cls.values()
 
 
-@pydantic.dataclasses.dataclass
-class ExportHandler(abc.ABC):
-    directory_path: pathlib.Path
+def _format_dict(data: dict[str, PrimitiveType]) -> dict[str, cabc.Sequence[PrimitiveType]]:
+    """Convert dict with heterogenous lengths of value to dict with where all values are lists of the same length."""
+    max_length = max(len(v) if isinstance(v, list) else 1 for v in data.values())
+    # Convert (non-list) values to lists with uniform length and fill (if necessary) shorter lists with None
+    return {
+        key: (v + [None] * (max_length - len(v))) if isinstance(v, list) else ([v] + [None] * (max_length - 1))
+        for key, v in data.items()
+    }
+
+
+class BaseIoHandler:
+    @staticmethod
+    def create_file_path(
+        *,
+        root_path: pathlib.Path,
+        file_type: FileType,
+        file_name: str | None = None,
+        study_case_name: str | None = None,
+    ) -> pathlib.Path:
+        timestamp = dt.datetime.now().astimezone()
+        timestamp_string = timestamp.isoformat(sep="T", timespec="seconds").replace(":", "")
+        study_case_name = f"{study_case_name}{NAME_SEPARATOR}" if study_case_name is not None else ""
+        filename = (
+            f"{study_case_name}{timestamp_string}{file_type.value}"
+            if file_name is None
+            else f"{study_case_name}{file_name}{file_type.value}"
+        )
+        file_path = root_path / filename
+        # Formal validation of path
+        try:
+            file_path.resolve()
+        except OSError as e:
+            msg = f"File path {file_path} is not a valid path."
+            raise FileNotFoundError(msg) from e
+
+        # Create (sub)directories if not existing
+        file_path.parent.mkdir(parents=True, exist_ok=True)
+
+        return file_path
+
+
+class PandasIoHandler(BaseIoHandler):
+    """Handler for importing and exporting data using pandas."""
+
+    try:
+        import pandas as pd  # noqa: PLC0415
+    except ModuleNotFoundError:
[…]
