+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 3.2.0 (2025-08-04)
+
+### Feat
+
+- handle dataframe for data export/import functions (#312)
+- add run_short_circuit_calculation() for pf version 2024 (#307)
+- added powerfactory_ini_name, powerfactory_user_password, and plausibility_check parameters to exporter interfaces and methods (#310)
+- add interface function get_slack_nodes() (#293)
+
+### Fix
+
+- add GetAttribute() to general pfi methods (#292)
+
 ## 3.1.0 (2025-02-06)
 
 ### Feat
