+++ pyproject.toml
@@ -17,7 +17,7 @@
   name = "ieeh-powerfactory-tools"
   readme = "README.md"
   requires-python = ">=3.10,<3.13"
-  version = "3.1.0"
+  version = "3.2.0"
 
   [project.optional-dependencies]
     io-pandas   = ["pandas", "pyarrow"]
@@ -113,7 +113,7 @@
 [tool.commitizen]
   name = "cz_conventional_commits"
   tag_format = "$version"
-  version = "3.1.0"
+  version = "3.2.0"
   version_files = [
     ".zenodo.json:version",
     "CITATION.cff:cff-version",
