### 2025-08-04 59406a2
fix: update readme and prepare new pypi release (#314)
Fixes #313
Fixes #315
