### 2025-08-26 5bdb69b
fix: pydantic validation error when creating meta data for export (#317)
Fixes #316
