### 2025-08-26 a76f2c3
bump: version 3.2.1 → 3.2.2
