+++ src/powerfactory_tools/__version__.py
@@ -3,4 +3,4 @@
 # :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2025.
 # :license: BSD 3-Clause
 
-VERSION = "3.2.1"
+VERSION = "3.2.2"
