### 2025-10-05 #8068 @robinroche
> > [@kyleniemeyer](https://github.com/kyleniemeyer): before answering, I need to make sure I have the required software. [@SebastianDD](https://github.com/SebastianDD): what are the requirements on the PowerFactory side? I only have an educational version, which does not include scripting features, as far as I know.
>
> Hi Robin [@robinroche](https://github.com/robinroche) , is there anything I can further do to support/fasten up the review process? BR

Hi, sorry for taking so long to reply. Way too many things on my list...
I only have access to PF 2023 through a network licence. I tried copying the 2024 folder and modying 2024 occurrences to 2023 in the folder. But I'm getting errors when running the control basic notebook:

```
ValidationError                           Traceback (most recent call last)
Cell In[6], [line 9](vscode-notebook-cell:?execution_count=6&line=9)
      6 import pathlib
      8 from powerfactory_tools.utils.io import FileType
----> [9](vscode-notebook-cell:?execution_count=6&line=9) from powerfactory_tools.versions.pf2023 import PowerFactoryInterface
     10 from powerfactory_tools.versions.pf2023.constants import NAME_SEPARATOR
     11 from powerfactory_tools.versions.pf2023.interface import ValidPythonVersion

File ~\AppData\Local\Packages\PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0\LocalCache\local-packages\Python311\site-packages\powerfactory_tools\versions\pf2023\__init__.py:5
      1 # :author: Sasan Jacob Rasti <sasan_jacob.rasti@tu-dresden.de>
      2 # :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-202[5](https://file+.vscode-resource.vscode-cdn.net/[...]~/AppData/Local/Packages/PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0/LocalCache/local-packages/Python311/site-packages/powerfactory_tools/versions/pf2023/__init__.py:5).
      3 # :license: BSD 3-Clause
----> 5 from powerfactory_tools.versions.pf2023.exporter.exporter import PowerFactoryExporter
      6 from powerfactory_tools.versions.pf2023.exporter.exporter import export_powerfactory_data
      7 from powerfactory_tools.versions.pf2023.interface import PowerFactoryInterface

File ~\AppData\Local\Packages\PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0\LocalCache\local-packages\Python311\site-packages\powerfactory_tools\versions\pf2023\exporter\exporter.py:78
     76 from powerfactory_tools.versions.pf2023.exporter.load_power import ControlTypeFactory
     77 from powerfactory_tools.versions.pf2023.exporter.load_power import LoadPower
---> [78](https://file+.vscode-resource.vscode-cdn.net/[...]/~/AppData/Local/Packages/PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0/LocalCache/local-packages/Python311/site-packages/powerfactory_tools/versions/pf2023/exporter/exporter.py:78) from powerfactory_tools.versions.pf2023.interface import DEFAULT_POWERFACTORY_PATH
     79 from powerfactory_tools.versions.pf2023.interface import DEFAULT_PYTHON_VERSION
     80 from powerfactory_tools.versions.pf2023.interface import POWERFACTORY_VERSION
```
