### 2025-10-06 #8068 @SebastianDD
> Hi, sorry for taking so long to reply. Way too many things on my list... I only have access to PF 2023 through a network licence. I tried copying the 2024 folder and modying 2024 occurrences to 2023 in the folder. But I'm getting errors when running the control basic notebook:
>
> ```
> ---------------------------------------------------------------------------
> ValidationError                           Traceback (most recent call last)
> Cell In[6], [line 9](vscode-notebook-cell:?execution_count=6&line=9)
>       6 import pathlib
>       8 from powerfactory_tools.utils.io import FileType
> ----> [9](vscode-notebook-cell:?execution_count=6&line=9) from powerfactory_tools.versions.pf2023 import PowerFactoryInterface
>      10 from powerfactory_tools.versions.pf2023.constants import NAME_SEPARATOR
>      11 from powerfactory_tools.versions.pf2023.interface import ValidPythonVersion
>
> File ~\AppData\Local\Packages\PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0\LocalCache\local-packages\Python311\site-packages\powerfactory_tools\versions\pf2023\__init__.py:5
>       1 # :author: Sasan Jacob Rasti <sasan_jacob.rasti@tu-dresden.de>
>       2 # :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-202[5](https://file+.vscode-resource.vscode-cdn.net/[...]~/AppData/Local/Packages/PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0/LocalCache/local-packages/Python311/site-packages/powerfactory_tools/versions/pf2023/__init__.py:5).
>       3 # :license: BSD 3-Clause
> ----> 5 from powerfactory_tools.versions.pf2023.exporter.exporter import PowerFactoryExporter
>       6 from powerfactory_tools.versions.pf2023.exporter.exporter import export_powerfactory_data
>       7 from powerfactory_tools.versions.pf2023.interface import PowerFactoryInterface
>
> File ~\AppData\Local\Packages\PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0\LocalCache\local-packages\Python311\site-packages\powerfactory_tools\versions\pf2023\exporter\exporter.py:78
>      76 from powerfactory_tools.versions.pf2023.exporter.load_power import ControlTypeFactory
>      77 from powerfactory_tools.versions.pf2023.exporter.load_power import LoadPower
> ---> [78](https://file+.vscode-resource.vscode-cdn.net/[...]/~/AppData/Local/Packages/PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0/LocalCache/local-packages/Python311/site-packages/powerfactory_tools/versions/pf2023/exporter/exporter.py:78) from powerfactory_tools.versions.pf2023.interface import DEFAULT_POWERFACTORY_PATH
>      79 from powerfactory_tools.versions.pf2023.interface import DEFAULT_PYTHON_VERSION
>      80 from powerfactory_tools.versions.pf2023.interface import POWERFACTORY_VERSION
> ```

Hi Robin, nevermind, I'm still appreciating your will for review.
So far so good, but I think the remaining error is sensitive to the interface.py line 67-69. It is mandatory that the `DEFAULT_POWERFACTORY_PATH` points to the directory, where your PowerFactory program is installed (regardless how the license is provided) --> see attached picture as example

FYI: the PowerFactory installation comes with a builtin Python module `*.pyd` which is used for communication. Within the `PowerFactoryInterface` the path of this module is built out of this constants:

```
interface.py line 172
module_path = (
            self.powerfactory_path / POWERFACTORY_VERSION / "Python" / self.python_version.value
            if self.powerfactory_service_pack is None else
            self.powerfactory_path / (POWERFACTORY_VERSION + f" SP{self.powerfactory_service_pack}") / "Python" / self.python_version.value
        )
```

I hope this resolves your error.
