### 2025-10-07 #8068 @SebastianDD
> [@SebastianDD](https://github.com/SebastianDD) if they can add some additional personal details (name, affiliation, link to webpage, etc.) to their github profiles, that would help

It's done. Please let me know if there are still open ends.
