### 2025-10-07 #8068 @robinroche
> Hi Robin, nevermind, I'm still appreciating your will for review. So far so good, but I think the remaining error is sensitive to the interface.py line 67-69. It is mandatory that the `DEFAULT_POWERFACTORY_PATH` points to the directory, where your PowerFactory program is installed (regardless how the license is provided) --> see attached picture as example
>
> FYI: the PowerFactory installation comes with a builtin Python module `*.pyd` which is used for communication. Within the `PowerFactoryInterface` the path of this module is built out of this constants:
>
> ```
> interface.py line 172
> module_path = (
>             self.powerfactory_path / POWERFACTORY_VERSION / "Python" / self.python_version.value
>             if self.powerfactory_service_pack is None else
>             self.powerfactory_path / (POWERFACTORY_VERSION + f" SP{self.powerfactory_service_pack}") / "Python" / self.python_version.value
>         )
> ```
>
> I hope this resolves your error.

Thanks for the help. I still can't make it work. I have PowerFactory 2023 SP4A and the following config:
- In the notebook:
```
# PowerFactory configuration
PF_SERVICE_PACK = "4A"  # mandatory
PF_USER_PROFILE = ""  # specification may be necessary
PF_INI_NAME = ""  # optional specification of ini file name to switch to full version (e.g. PowerFactoryFull for file PowerFactoryFull.ini)
PF_PYTHON_VERSION = ValidPythonVersion.VERSION_3_11  # python version of local code environment must match the python version of PowerFactory API
```
- In interface.py:
```
POWERFACTORY_VERSION = "PowerFactory 2023"
DEFAULT_POWERFACTORY_PATH = pathlib.Path("C:/Program Files/DIgSILENT")
DEFAULT_PYTHON_VERSION = ValidPythonVersion.VERSION_3_11
```
- This error:
```
ValidationError                           Traceback (most recent call last)
Cell In[2], line 9
      6 import pathlib
      8 from powerfactory_tools.utils.io import FileType
----> 9 from powerfactory_tools.versions.pf2023 import PowerFactoryInterface
     10 from powerfactory_tools.versions.pf2023.constants import NAME_SEPARATOR
     11 from powerfactory_tools.versions.pf2023.interface import ValidPythonVersion

File ~\AppData\Local\Packages\PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0\LocalCache\local-packages\Python311\site-packages\powerfactory_tools\versions\pf2023\__init__.py:5
      1 # :author: Sasan Jacob Rasti <sasan_jacob.rasti@tu-dresden.de>
      2 # :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2025.
      3 # :license: BSD 3-Clause
----> 5 from powerfactory_tools.versions.pf2023.exporter.exporter import PowerFactoryExporter
      6 from powerfactory_tools.versions.pf2023.exporter.exporter import export_powerfactory_data
      7 from powerfactory_tools.versions.pf2023.interface import PowerFactoryInterface

File ~\AppData\Local\Packages\PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0\LocalCache\local-packages\Python311\site-packages\powerfactory_tools\versions\pf2023\exporter\exporter.py:78
     76 from powerfactory_tools.versions.pf2023.exporter.load_power import ControlTypeFactory
     77 from powerfactory_tools.versions.pf2023.exporter.load_power import LoadPower
---> 78 from powerfactory_tools.versions.pf2023.interface import DEFAULT_POWERFACTORY_PATH
     79 from powerfactory_tools.versions.pf2023.interface import DEFAULT_PYTHON_VERSION
     80 from powerfactory_tools.versions.pf2023.interface import POWERFACTORY_VERSION

File ~\AppData\Local\Packages\PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0\LocalCache\local-packages\Python311\site-packages\powerfactory_tools\versions\pf2023\interface.py:95
     91     cspqexpgen: MetricPrefix  # Generators etc.
     92     currency: Currency
---> 95 DEFAULT_PROJECT_UNIT_SETTING = ProjectUnitSetting(
     96     ilenunit=UnitSystem.METRIC,
     97     clenexp=BaseUnits.LENGTH,
     98     cspqexp=BaseUnits.POWER,
     99     cspqexpgen=BaseUnits.POWER,
    100     currency=BaseUnits.CURRENCY,
    101 )
    104 @pydantic.dataclasses.dataclass
    105 class PowerFactoryInterface:
    106     project_name: str

File ~\AppData\Local\Packages\PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0\LocalCache\local-packages\Python311\site-packages\pydantic\_internal\_dataclasses.py:120, in complete_dataclass.<locals>.__init__(__dataclass_self__, *args, **kwargs)
    118 __tracebackhide__ = True
    119 s = __dataclass_self__
--> 120 s.__pydantic_validator__.validate_python(ArgsKwargs(args, kwargs), self_instance=s)

ValidationError: 4 validation errors for ProjectUnitSetting
clenexp
  Input should be 'a', 'f', 'p', 'n', 'u', 'm', '', 'k', 'M', 'G', 'T', 'P' or 'E' [type=enum, input_value=<MetricPrefix.k: 'k'>, input_type=MetricPrefix]
    For further information visit https://errors.pydantic.dev/2.11/v/enum
cspqexp
  Input should be 'a', 'f', 'p', 'n', 'u', 'm', '', 'k', 'M', 'G', 'T', 'P' or 'E' [type=enum, input_value=<MetricPrefix.M: 'M'>, input_type=MetricPrefix]
    For further information visit https://errors.pydantic.dev/2.11/v/enum
cspqexpgen
  Input should be 'a', 'f', 'p', 'n', 'u', 'm', '', 'k', 'M', 'G', 'T', 'P' or 'E' [type=enum, input_value=<MetricPrefix.M: 'M'>, input_type=MetricPrefix]
    For further information visit https://errors.pydantic.dev/2.11/v/enum
currency
  Input should be 'USD', 'EUR', 'JPY', 'GBP', 'AUD', 'CAD', 'CHF', 'CNY', 'SEK', 'MXN', 'NZD', 'SGD', 'HKD', 'NOK', 'KRW', 'TRY', 'INR', 'RUB', 'BRL', 'ZAR' or 'CLP' [type=enum, input_value=<Currency.EUR: 'EUR'>, input_type=Currency]
    For further information visit https://errors.pydantic.dev/2.11/v/enum
```
Any idea?
