### 2025-10-08 #8068 @SebastianDD
@robinroche well that looks strange. But now I understand that both errors of yours are the same. This errors are unrelated to PowerFactory, but aims on the internal data handling and validation using `pydantic`. --> Maybe looks like a pydantic bug.
I did some tests: I also installed a Python 3.11.(3.11.3 for me) and run the first section of the "powerfactory_control__basic" notebook. No error occurs on my system for all pydantic versions 2.11.0 - 2.11.10.

For further debugging: Could you please provide the python version you used as well as the installed libs in your virtual environment? Relevant for me is the version of `powerfactory-tools` and `pydantic`.
Thanks in advance.
