### 2025-10-17 #8068 @DorotheeNitsch
Dear @kyleniemeyer,

Thank you for considering me as a reviewer for this submission. I’ve checked the functionality on my system and have access to DigSilent PowerFactory, as well as experience with it. I was able to install PowerFactory-Tools and run the first example successfully, so I’m happy to accept the review.
