### 2025-10-21 7fc53a9
build(pypi): add native Powerfactory 2025 and Python 3.13 support Fixes #318
