+++ src/powerfactory_tools/versions/pf2025/exporter/exporter.py
@@ -1,0 +1,4286 @@
+# :author: Sasan Jacob Rasti <sasan_jacob.rasti@tu-dresden.de>
+# :author: Sebastian Krahmer <sebastian.krahmer@tu-dresden.de>
+# :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2025.
+# :license: BSD 3-Clause
+
+from __future__ import annotations
+
+import datetime as dt
+import itertools
+import logging
+import math
+import multiprocessing
+import pathlib
+import textwrap
+import typing as t
+from collections.abc import Sequence
+from sys import setrecursionlimit
+
+import loguru
+import pydantic
+from psdm.base import AttributeData
+from psdm.base import UniqueTuple
+from psdm.base import VoltageSystemType
+from psdm.meta import Meta
+from psdm.meta import SignConvention
+from psdm.quantities.multi_phase import Phase
+from psdm.quantities.multi_phase import PowerFactorDirection
+from psdm.quantities.single_phase import AdmittanceNat
+from psdm.quantities.single_phase import AdmittancePosSeq
+from psdm.quantities.single_phase import AdmittanceZerSeq
+from psdm.quantities.single_phase import ImpedanceNat
+from psdm.quantities.single_phase import ImpedancePosSeq
+from psdm.quantities.single_phase import ImpedanceZerSeq
+from psdm.quantities.single_phase import Length
+from psdm.quantities.single_phase import PhaseAngleClock
+from psdm.steadystate_case.active_power import ActivePower as ActivePowerSSC
+from psdm.steadystate_case.case import Case as SteadystateCase
+from psdm.steadystate_case.characteristic import Characteristic
+from psdm.steadystate_case.controller import ControlledVoltageRef
+from psdm.steadystate_case.controller import ControlQP
+from psdm.steadystate_case.controller import PController
+from psdm.steadystate_case.controller import QController
+from psdm.steadystate_case.controller import QControlStrategy
+from psdm.steadystate_case.external_grid import ExternalGrid as ExternalGridSSC
+from psdm.steadystate_case.load import Load as LoadSSC
+from psdm.steadystate_case.reactive_power import ReactivePower as ReactivePowerSSC
+from psdm.steadystate_case.transformer import Transformer as TransformerSSC
+from psdm.topology.branch import Branch
+from psdm.topology.branch import BranchType
+from psdm.topology.external_grid import ExternalGrid
+from psdm.topology.external_grid import GridType
+from psdm.topology.load import Load
+from psdm.topology.load import LoadType
+from psdm.topology.load import PhaseConnections
+from psdm.topology.load import SystemType
+from psdm.topology.load_model import LoadModel
+from psdm.topology.node import Node
+from psdm.topology.topology import Topology
+from psdm.topology.transformer import TapSide
+from psdm.topology.transformer import Transformer
+from psdm.topology.transformer import TransformerPhaseTechnologyType
+from psdm.topology.transformer import VectorGroup as TVectorGroup
+from psdm.topology.windings import VectorGroup as WVectorGroup
+from psdm.topology.windings import Winding
+from psdm.topology_case.case import Case as TopologyCase
+from psdm.topology_case.element_state import ElementState
+
+from powerfactory_tools.__version__ import VERSION
+from powerfactory_tools.str_constants import NAME_SEPARATOR
+from powerfactory_tools.str_constants import STRING_SEPARATOR
+from powerfactory_tools.utils.io import FileType
+from powerfactory_tools.versions.pf2025.constants import DEFAULT_PHASE_QUANTITY
+from powerfactory_tools.versions.pf2025.constants import DecimalDigits
[…]
