### 2025-10-21 #9281 @editorialbot
[REVIEW]: PowerFactory-Tools: A toolbox to facilitate the control of DIgSILENT PowerFactory, written in Python
**Submitting author:** @SebastianDD (<a href="http://orcid.org/0000-0002-3940-3398">Sebastian Krahmer</a>)
**Repository:** https://github.com/ieeh-tu-dresden/powerfactory-tools
**Branch with paper.md** (empty if default branch): SebastianDD/166-docs-add-JOSS-paper
**Version:** 3.3.1
