### 2025-10-27 1af6e0b
fix: update python versions in pyproject.toml
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
