+++ pyproject.toml
@@ -15,7 +15,7 @@
   license = { file = "LICENSE" }
   name = "ieeh-powerfactory-tools"
   readme = "README.md"
-  requires-python = ">3.10,<3.14"
+  requires-python = ">=3.10,<3.14"
   version = "3.2.2"
 
   [project.optional-dependencies]
