### 2025-10-28 594225f
build(pypi): add native Powerfactory 2025 and Python 3.13 support (#321)
Fixes #318
