+++ src/powerfactory_tools/versions/pf2025/exporter/load_power.py
@@ -1,0 +1,881 @@
+# :author: Sasan Jacob Rasti <sasan_jacob.rasti@tu-dresden.de>
+# :author: Sebastian Krahmer <sebastian.krahmer@tu-dresden.de>
+# :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2025.
+# :license: BSD 3-Clause
+
+from __future__ import annotations
+
+import enum
+import itertools
+import math
+import typing as t
+from dataclasses import dataclass
+
+import loguru
+from psdm.quantities.multi_phase import ActivePower
+from psdm.quantities.multi_phase import ApparentPower
+from psdm.quantities.multi_phase import CosPhi
+from psdm.quantities.multi_phase import ReactivePower
+from psdm.quantities.multi_phase import TanPhi
+from psdm.quantities.single_phase import PowerFactorDirection
+from psdm.quantities.single_phase import SystemType
+from psdm.steadystate_case.characteristic import Characteristic
+from psdm.steadystate_case.controller import ControlCosPhiConst
+from psdm.steadystate_case.controller import ControlCosPhiP
+from psdm.steadystate_case.controller import ControlCosPhiU
+from psdm.steadystate_case.controller import ControlledVoltageRef
+from psdm.steadystate_case.controller import ControlPConst
+from psdm.steadystate_case.controller import ControlQConst
+from psdm.steadystate_case.controller import ControlQP
+from psdm.steadystate_case.controller import ControlQU
+from psdm.steadystate_case.controller import ControlTanPhiConst
+from psdm.steadystate_case.controller import ControlUConst
+from psdm.steadystate_case.controller import QControlStrategy
+from psdm.topology.load import RatedPower
+
+from powerfactory_tools.versions.pf2025.constants import DecimalDigits
+from powerfactory_tools.versions.pf2025.constants import Exponents
+from powerfactory_tools.versions.pf2025.quantities import QuantityConverter as Qc
+
+if t.TYPE_CHECKING:
+    from typing import TypedDict
+
+    class PowerDict(TypedDict):
+        power_apparent: float
+        power_active: float
+        power_reactive: float
+        cos_phi: float
+        power_factor_direction: PowerFactorDirection
+        power_reactive_control_type: QControlStrategy
+
+
+class ConsolidatedLoadPhaseConnectionType(enum.Enum):
+    ONE_PH_PH_E = "ONE_PH_PH_E"
+    ONE_PH_PH_N = "ONE_PH_PH_N"
+    ONE_PH_PH_PH = "ONE_PH_PH_PH"
+    THREE_PH_D = "THREE_PH_D"
+    THREE_PH_PH_E = "THREE_PH_PH_E"
+    THREE_PH_Y = "THREE_PH_Y"
+    THREE_PH_YN = "THREE_PH_YN"
+    TWO_PH_PH_E = "TWO_PH_PH_E"
+    TWO_PH_Y = "TWO_PH_Y"
+    TWO_PH_YN = "TWO_PH_YN"
+
+
+@dataclass
+class LoadPower:
+    pow_apps: tuple[float, ...]
+    pow_acts: tuple[float, ...]
+    pow_reacts: tuple[float, ...]
+    cos_phis: tuple[float, ...]
+    pow_fac_dir: PowerFactorDirection
+    pow_react_control_type: QControlStrategy
+
+    @staticmethod
+    def _is_symmetrical(values: tuple[float, ...] | tuple[str, ...]) -> bool:
+        return len(list(itertools.groupby(values))) in (0, 1)
+
+    @property
+    def pow_app(self) -> float:
+        return sum(self.pow_apps)
+
+    @property
[…]
