+++ src/powerfactory_tools/versions/pf2025/types.py
@@ -1,0 +1,1957 @@
+# :author: Sasan Jacob Rasti <sasan_jacob.rasti@tu-dresden.de>
+# :author: Sebastian Krahmer <sebastian.krahmer@tu-dresden.de>
+# :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2025.
+# :license: BSD 3-Clause
+
+from __future__ import annotations
+
+import enum
+import typing as t
+from typing import Protocol
+
+if t.TYPE_CHECKING:
+    from collections.abc import Sequence
+
+
+# High-Level Enums
+class PFClassId(enum.Enum):
+    AREA = "ElmArea"
+    COMPOUND_GRID_ELEMENT = "ElmFolder"  # e.g. a compound grid graphic consisting of multiple elements
+    COMPOUND_MODEL = "ElmComp"  # e.g. a compound generator with multiple functional slots as part of a template model
+    COUPLER = "ElmCoup"
+    CUBICLE = "StaCubic"
+    CURRENT_SOURCE_AC = "ElmIac"
+    EXTERNAL_GRID = "ElmXnet"
+    DATETIME = "SetTime"
+    DESKTOP = "SetDesktop"
+    DSL_MODEL = "ElmDsl"
+    FOLDER = "IntFolder"
+    FUSE = "RelFuse"
+    FUSE_TYPE = "TypFuse"
+    GENERATOR = "ElmGenstat"
+    GRID = "ElmNet"
+    GRID_GRAPHIC = "IntGrfnet"
+    LINE = "ElmLne"
+    LINE_TYPE = "TypLne"
+    LOAD = "ElmLod"
+    LOAD_FLOW_CONTROLLER = "ElmLdfctrl"
+    LOAD_LV = "ElmLodlv"
+    LOAD_LV_PART = "ElmLodlvp"
+    LOAD_MV = "ElmLodmv"
+    LOAD_TYPE_GENERAL = "TypLod"
+    LOAD_TYPE_LV = "TypLodlv"
+    LOAD_TYPE_MV = "TypDistrf"
+    MEASUREMENT_FILE = "ElmFile"
+    PROJECT_FOLDER = "IntPrjfolder"
+    PROJECT_SETTINGS = "SetPrj"
+    PVSYSTEM = "ElmPvsys"
+    REFERENCE = "IntRef"
+    RESULT = "ElmRes"
+    SCENARIO = "IntScenario"
+    SCRIPT_DPL = "ComDpl"
+    SCRIPT_PYTHON = "ComPython"
+    SELECTION = "SetSelect"
+    SETTINGS_FOLDER = "SetFold"
+    SETTINGS_FOLDER_UNITS = "IntUnit"
+    SHUNT = "ElmShnt"
+    SOURCE_TYPE_HARMONIC_CURRENT = "TypHmccur"
+    STATION_CONTROLLER = "ElmStactrl"
+    STUDY_CASE = "IntCase"
+    SUBSTATION = "ElmSubstat"
+    SUBSTATION_FIELD = "ElmBay"
+    SWITCH = "StaSwitch"
+    TEMPLATE = "IntTemplate"
+    TERMINAL = "ElmTerm"
+    TRANSFORMER_2W = "ElmTr2"
+    TRANSFORMER_2W_TYPE = "TypTr2"
+    TRANSFORMER_3W = "ElmTr3"
+    TRANSFORMER_3W_TYPE = "TypTr3"
+    UNIT_VARIABLE = "SetVariable"
+    VARIABLE_MONITOR = "IntMon"  # Variable monitor definition
+    VARIANT = "IntScheme"
[…]
