+++ .github/workflows/create-release.yml
@@ -14,11 +14,11 @@ jobs:
     runs-on: ubuntu-latest
     needs: [test-code]
     env:
-      PYTHON_VERSION: "3.10"
+      PYTHON_VERSION: "3.12"
 
     steps:
       - name: Check out repository
-        uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
+        uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8  # v5.0.0
         with:
           token: ${{ secrets.ACCESS_TOKEN }}
           fetch-depth: 0
@@ -44,7 +44,7 @@ jobs:
 
       - name: Find new version
         id: cz
-        uses: commitizen-tools/commitizen-action@a49cac8e98f3954a16e107f026f71cf84b17706c  # 0.23.1
+        uses: commitizen-tools/commitizen-action@5b0848cd060263e24602d1eba03710e056ef7711  # 0.24.0
         with:
           github_token: ${{ secrets.ACCESS_TOKEN }}
           push: false
@@ -52,7 +52,7 @@ jobs:
           changelog: false
 
       - name: Check out repository
-        uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
+        uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8  # v5.0.0
         with:
           token: ${{ secrets.ACCESS_TOKEN }}
           fetch-depth: 0
@@ -66,9 +66,9 @@ jobs:
 
       - name: Install uv and set the python version
         id: setup-uv
-        uses: astral-sh/setup-uv@v5
+        uses: astral-sh/setup-uv@v7
         with:
-          version: "0.5.28"  # install a specific version of uv
+          version: "0.9.4"  # install a specific version of uv
           enable-cache: true
           python-version: ${{ env.PYTHON_VERSION  }}
 
@@ -111,7 +111,7 @@ jobs:
         run: echo "`jq '.related_identifiers[0].identifier="https://github.com/ieeh-tu-dresden/powerfactory-tools/tree/${{ steps.cz.outputs.version }}"' .zenodo.json`" > .zenodo.json
 
       - name: Bump version and create changelog
-        uses: commitizen-tools/commitizen-action@a49cac8e98f3954a16e107f026f71cf84b17706c  # 0.23.1
+        uses: commitizen-tools/commitizen-action@5b0848cd060263e24602d1eba03710e056ef7711  # 0.24.0
         id: commitizen
         with:
           github_token: ${{ secrets.ACCESS_TOKEN }}
@@ -134,7 +134,7 @@ jobs:
           UV_PUBLISH_TOKEN: ${{ secrets.PYPI_TOKEN }}
 
       - name: Create release
-        uses: softprops/action-gh-release@c95fe1489396fe8a9eb87c0abf8aa5b2ef267fda  # v2.2.1
+        uses: softprops/action-gh-release@6da8fa9354ddfdc4aeace5fc48d7f679b5214090  # v2.4.1
         with:
           body_path: body.md
           tag_name: ${{ steps.commitizen.outputs.version }}
