### 2025-10-30 fac5dbe
feat: support PF2025 and Python 3.13 (#323)
Fixes #322
