+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.3.0 (2025-10-30)
+
+### Feat
+
+- support PF2025 and Python 3.13 (#323)
+
 ## 3.2.2 (2025-08-26)
 
 ### Fix
