+++ CITATION.cff
@@ -1,4 +1,4 @@
-cff-version: 3.2.2
+cff-version: 3.3.0
 message: "If you use this software, please cite it as below."
 authors:
   - family-names: "Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden"
