### 2025-11-11 6c9f14e
docs: implement reviewer bugfixes in JOSS paper Fixes #166
