### 2025-11-11 #9281 @SebastianDD
Fixed code snippet provided in paper.
