### 2025-11-14 b47c52f
docs: fix placeholder in joss paper code to be in line with jupyter examples
