### 2025-11-17 #9281 @SebastianPalmDD
**Conflict of interest**
I worked at the same chair at TU Dresden as Sebastian Krahmer, Sasan, Maximilian and Laura. The PowerFactory Tools, however, have been developed after I left the university, therefore I do not see this as a conflict of interest.

In the following, I will point out some specific checkpoints from the review list. For everything else the checked boxes should be self-explanatory.

**General Checks
Reproducibility:**
With the update of the tool, which now supports PowerFactory 2025 and Python 3.13, I was able to reproduce the examples provided after some initial difficulties.

**Documentation
Automated Tests:**
As far as I have seen, there are no automated tests. Ultimately, manual verification of the functions consists of comparing the JSON/PDSM files with the PowerFactory project (for the import/export functionality) or performing the calculations manually to ensure that the results match. I do not see a feasible way to automate this.

**Software Paper**
The paper is well-written and contains all the information necessary to understand the purpose and use of the tool.
I could only find on minor language mistake where on the first page “It has to mention” should be replaced with “It has to be mentioned”.
