### 2025-11-25 bee574f
docs: clarify description of powerfactory in paper.md
