### 2025-11-25 9520372
docs: fix links in both readmes
