### 2025-12-03 #9281 @SebastianDD
@kyleniemeyer
> please update the author list of the archive to match the author list of the paper - you can specify the affiliation for each person

Done.

> my only comment about the paper itself is that there appears to be a bulletedlist (or what was intended to be a bulleted list) in the "Power System Data Model" paragraph - did you intend to format that as a list

Yes, this was intended.
In locally compiled markdown it looks like:

> can be stored as a human-readable JSON file:
> - Topology: plain network model with nodes, edges and connected devices
> - TopologyCase: information about elements that are disconnected, e. g. out-of-service or via open switches
> - SteadystateCase: operational case specific information.

Do you have any suggestion to fix this?
