### 2025-12-03 f07a5d5
docs(fix): adjust referred Zenodo reference
