+++ docs/joss_paper/paper.bib
@@ -32,7 +32,7 @@
     label = {PowerFactory-Tools},
     doi = {10.5281/zenodo.7074968},
     license = {BSD 3-Clause},
-    title = {{PowerFactory Tools - A toolbox for Python based control of DIgSILENT PowerFactory}},
+    title = {{PowerFactory-Tools: A Python Package to Facilitate the Control of DIgSILENT PowerFactory}},
     url = {https://github.com/ieeh-tu-dresden/powerfactory-tools},
     year = {2022},
     howpublished={GitHub Repository}
