### 2025-12-03 7119b47
docs(fix): fix year of used pf version, set to 2025
