### 2025-12-03 #9281 @kyleniemeyer
@SebastianDD please take a look at the paper proof and confirm that it is satisfactory
