### 2025-12-04 #9281 @SebastianDD
> [@SebastianDD](https://github.com/SebastianDD) please take a look at the paper proof and confirm that it is satisfactory

@kyleniemeyer I confirm that the final proof is satisfactory.

Thank you for hosting the editorial process. 🚀
